pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hopper"
version = "0.1.0"
description = "Synthetic CRES signal + track simulator (cyclotron + axial + grad-B drift)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "E. Karim"}]
keywords = ["CRES", "Project8", "simulation", "cyclotron"]
dependencies = [
  "numpy>=1.23",
  "scipy>=1.9",
  "pyyaml>=6.0",
  "h5py>=3.8"
]

[project.optional-dependencies]
root = ["uproot>=5.0"]

[project.scripts]
hopper-sim = "hopper.cli.run_sim:main"

[tool.setuptools.packages.find]
where = ["src"]
